docs(linux): J742S2: Update the USB documentation
All the J7 family SoCs share same USB documentation. This documentation
is incorrect in the case of J742S2 EVM. Update the documentation for
J742S2 EVM by adding 'J742S2' as a "CONFIG_part_variant" in the USB
documentation for J784S4 EVM as same documentation is valid for both
the platforms.

Signed-off-by: Hrushikesh Salunke <h-salunke@ti.com>